+++ test/faker/app_test.exs
@@ -1,0 +1,15 @@
+defmodule AppTest do
+  use ExUnit.Case, async: true
+
+  test "author/0" do
+    assert is_binary(Faker.App.author)
+  end
+
+  test "name/0" do
+    assert is_binary(Faker.App.name)
+  end
+
+  test "version/0" do
+    assert is_binary(Faker.App.version)
+  end
+end